+++ cypress/integration/blocks.js
@@ -40,7 +40,7 @@ if (Cypress.env('API') !== 'guillotina') {
       cy.get('.ui.basic.icon.button.video')
         .contains('video')
         .click();
-      cy.get('.toolbar > .ui > input')
+      cy.get('.toolbar-inner > .ui > input')
         .click()
         .type('https://youtu.be/T6J3d35oIAY')
         .type('{enter}');
@@ -60,7 +60,7 @@ if (Cypress.env('API') !== 'guillotina') {
       cy.get('.ui.basic.icon.button.video')
         .contains('video')
         .click();
-      cy.get('.toolbar > .ui > input')
+      cy.get('.toolbar-inner > .ui > input')
         .click()
         .type('https://vimeo.com/85804536')
         .type('{enter}');